requirements.txt
flask
gunicorn
supabase
python-dotenv
requests
